WIP implementing thor commands
This version was developed with a ruby gem for an older version of the
Linode API.  That API has been deprecated, and there is currently no
updated version of the gem.  Accordingly, for now I'm halting
development of this version and will focus on developing the Bash shell
variant of this utility (see joelhelbling/linode-shell).